use the same vit in both pretrain and linear prob
