+++ projects/MOCOV3/configs/models/vit_base_patch16.py
@@ -1,4 +1,3 @@
-from torch import is_inference
 from libai.config import LazyCall
 from modeling.vit import VisionTransformer
 
@@ -12,6 +11,4 @@ model = LazyCall(VisionTransformer)(
         mlp_ratio=4,
         drop_path_rate=0.1,
         global_pool=False,
-        finetune=None,
-        weight_style=None
 )
