+++ projects/MOCOV3/modeling/vit.py
@@ -20,10 +20,19 @@
 # mae: https://github.com/facebookresearch/mae/blob/main/models_vit.py
 # --------------------------------------------------------
 
+
+import math
+
 import oneflow as flow
+import oneflow.nn as nn
+
 from flowvision.layers.weight_init import trunc_normal_
 
+from operator import mul
+from functools import reduce
+
 import libai.models.vision_transformer
+from libai.layers import Linear, PatchEmbedding
 from utils.weight_convert import load_torch_checkpoint_finetune
 
 
@@ -48,8 +57,9 @@ class VisionTransformer(libai.models.vision_transformer.VisionTransformer):
         global_pool=False,
         num_classes=1000,
         loss_func=None,
-        finetune=None,
-        weight_style="pytorch"
+        linear_prob=None,
+        weight_style="pytorch",
+        stop_grad_conv1=False
     ):
         super(VisionTransformer, self).__init__(
             img_size=img_size,
@@ -68,12 +78,64 @@ class VisionTransformer(libai.models.vision_transformer.VisionTransformer):
         self.global_pool = global_pool
 
         # weight init
-        if finetune:
-            self.load_checkpoint(finetune, weight_style, num_heads, embed_dim)
+        if linear_prob:
+            self.load_checkpoint(linear_prob, weight_style, num_heads, embed_dim)
+            self.head.weight.data.normal_(mean=0.0, std=0.01)
+            self.head.bias.data.zeros_()
         else:
             trunc_normal_(self.pos_embed, std=0.02)
             trunc_normal_(self.cls_token, std=0.02)
             self.apply(self._init_weights)
+
+            self.stop_grad_conv1 = stop_grad_conv1
+            self.embed_dim = embed_dim
+            self.initialization()
+
+    def initialization(self):
+
+        # Use fixed 2D sin-cos position embedding
+        self.build_2d_sincos_position_embedding()
+
+        # weight initialization
+        for name, m in self.named_modules():
+            if isinstance(m, Linear): 
+                if 'query_key_value' in name:
+                    val = math.sqrt(6. / float(m.weight.shape[0] // 3 + m.weight.shape[1])) 
+                    nn.init.uniform_(m.weight, -val, val)
+                else:
+                    nn.init.xavier_uniform_(m.weight)
+
+                nn.init.zeros_(m.bias)
+        nn.init.normal_(self.cls_token, std=1e-6)
+
+        if isinstance(self.patch_embed, PatchEmbedding):
+            # xavier_uniform initialization
+            val = math.sqrt(6. / float(3 * reduce(mul, self.patch_embed.patch_size, 1) + self.embed_dim))
+            nn.init.uniform_(self.patch_embed.proj.weight, -val, val)
+            nn.init.zeros_(self.patch_embed.proj.bias)
+
+            if self.stop_grad_conv1:
+                self.patch_embed.proj.weight.requires_grad = False
+                self.patch_embed.proj.bias.requires_grad = False
+
+    def build_2d_sincos_position_embedding(self, temperature=10000.):
+        sbp = self.pos_embed.sbp
+        placement = self.pos_embed.placement
+
+        h, w = self.patch_embed.grid_size
+        grid_w = flow.arange(w, dtype=flow.float32).cuda().to_global(sbp=sbp, placement=placement)
+        grid_h = flow.arange(h, dtype=flow.float32).cuda().to_global(sbp=sbp, placement=placement)
+        grid_w, grid_h = flow.meshgrid(grid_w, grid_h)
+        assert self.embed_dim % 4 == 0, 'Embed dimension must be divisible by 4 for 2D sin-cos position embedding'
+        pos_dim = self.embed_dim // 4
+        omega = (flow.arange(pos_dim, dtype=flow.float32) / pos_dim).cuda().to_global(sbp=sbp, placement=placement)
+        omega = (1. / flow.tensor(temperature).cuda().to_global(sbp=sbp, placement=placement)**omega)
+        out_w = flow.einsum('m,d->md', grid_w.flatten(), omega)
+        out_h = flow.einsum('m,d->md', grid_h.flatten(), omega)
+        pos_emb = flow.cat([flow.sin(out_w), flow.cos(out_w), flow.sin(out_h), flow.cos(out_h)], dim=1)[None, :, :]
+        pe_token = flow.zeros([1, 1, self.embed_dim], dtype=flow.float32).cuda().to_global(sbp=sbp, placement=placement)
+        self.pos_embed = nn.Parameter(flow.cat([pe_token, pos_emb], dim=1))
+        self.pos_embed.requires_grad = False
 
     def load_checkpoint(self, finetune, weight_style, num_heads, embed_dim):
         linear_keyword = "head"
@@ -85,9 +147,6 @@ class VisionTransformer(libai.models.vision_transformer.VisionTransformer):
             params = load_torch_checkpoint_finetune(num_heads, embed_dim, path=finetune)
         else:
             params = flow.load(finetune)
-
-        self.head.weight.data.normal_(mean=0.0, std=0.01)
-        self.head.bias.data.zeros_()
 
         self.load_state_dict(params, strict=False)
